[UPDATE]
Changing id's from Integer to Long

+++ adoptaMe/src/main/java/mx/edu/utez/adoptaMe/entity/Color.java
@@ -19,7 +19,7 @@ public class Color {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "name",nullable = false)
     private String name;
@@ -27,16 +27,16 @@ public class Color {
     public Color() {
     }
 
-    public Color(Integer id, String name) {
+    public Color(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
