+++ adoptaMe/src/main/java/mx/edu/utez/adoptaMe/entity/Personality.java
@@ -19,7 +19,7 @@ public class Personality {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "name", nullable = false)
     private String name;
@@ -27,16 +27,16 @@ public class Personality {
     public Personality() {
     }
 
-    public Personality(Integer id, String name) {
+    public Personality(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
